Add `doc` target to build D documentation
New `doc` target will, by default, call harbored-mod
tool to generate the documentation for the project from
DDOC comments inside source files.

Harbored-mod is choosen because it also allows Markdown
syntax which makes the documentation easier to read in
the source files, as it doesn't require as much DDOC macros
as the dmd.