git-deblog: Warn about wrong decoding
Instead of vomiting a backtrace to the user if decoding errors are
found, print a nice(r) warning saying X error were found while trying to
decode the output of a particular command, using a particular encoding.

The program will also continue working normally and the broken
characters will be output using a simple backslash replacement string.

This is an improvement over #30 and related to #33.