Move all unittest processing to the unittest target
This commit moves the compilation and running of unittests to the
`unittest` target(s) and remove the flags `-unittest -debug=UnitTest
-version=UnitTest` when building integrationtests.

To do this now we need to stop excluding `main.d` files from unittests,
as integration tests are usually just a `main.d` file, so if we exclude
those no unittests will be run for integration tests. For consistency we
stop automatically excluding project's `main.d` files too. Instead we
require projects to conditionally compile the `main()` function only
when `-version=UnitTest` is not present.

Also a INTEGRATIONTEST variable is added to change the default location
(`test/`) for integration tests. This will be backported to v1.x.x and
thus not added to the release notes.

Fixes #41.