test: Add exception for importing x.y.package
When generating the `allunitests.d` file, for every file an `import` is
generated based on the file name, but in D2 the file `pkg/package.d` is
also allowed, but is imported when `pkg` is imported

Since using `package` as a file/import name is invalid in D1, is safe to
make this exception regardless of the D version being used.

To get more consistency when converting paths to modules, a new function
(`file2module`) is added too.

Fixes #73.