Make version information more concise
Pass the `--abbrev` argument to `git describe` with a value of 4. This
makes the output of `git describe` more concise by using only 4 digits
of the commit hash if that is enough to uniquely identify a git object.
The number of digits used could be more if multiple git objects have the
same 4 digits in the beginning.

This also means that we have to adapt the regex that parses the `git
describe` output to match 4 or more digits in the commit hash.