Fix for replacing `@DFLAGS@` with build flags
This change replaces the sed delimiter used for replacing the string
`@DFLAGS@` with the contents of `$DFLAGS` to use `\#` instead of `/`.
This is safer because `#` is unlikely to be in `$DFLAGS` whereas `/` is
part of most filesystem paths. An alternative is to escape any usage of
`/` in the contents of `$DFLAGS`.

Fixes sociomantic-tsunami/makd#111.