mkversion: Fix replacing of info in Version.d.tpl
The replacing of build information in the Version.d.tpl is done via
a sed replace command (s/@PLACEHOLDER@/$value/), which can misbehave if
the value contain "special" characters (like "/"). This is error prone
and unneeded.

The replacement is now done via awk using gsub(), which performs a very
dumb literal search and replace, which is what we need here.

Fixes #111.