Use a new scheme for Debian versions
Debian versions used by MakD are broken when the git tag is
a pre-release as per SemVer. The problem is the `-` separator for SemVer
pre-releases is not translated so, for example, version 1.2.3-xenial is
older than 1.2.3-alpha.1-xenial, because by having 2 `-` in the version,
then the whole string is interpreted as the upstream version, and `-`
loses any special meaning.

To fix this the new scheme is used that should solve all the issues.
Also `mkversion.sh` is adjusted to print untagged versions as
<branch>-0-g<H> instead of the old <branch>-g<H>. This is to simplify
parsing when converting to Debian versions.

This new scheme is as follows:

The input version is taken from STDIN and is expected to be in the form:
[v]<upstream_version>[-<C>-g<H>][-dirty] (as constructed by
`mkversion.sh`).

Where:

<upstream_version>
  Is the current version of the project, expected to be the last tag, which
  should be SemVer compatible, or a branch name.
  It will be parsed as a SemVer version:
    <X>.<Y>.<Z>[-<pre_release>][+<metadata>]
   If the parsing is not successful, <X>, <Y> and <Z> are interpreted as 0 and
   the <pre_release> is set to `git.<upstream_version>` (so is interpreted as
   a 0.0.0 pre-release). In this case any non-alphanumeric character that is
   not compatible with the SemVer character set allowed for pre-releases is
   replaced by `.`.

  In any case `-` chars in <pre_release> and <metadata> is translated to `.`
  (as `-` has a special meaning for Debian versions).

<C>
  Is how many commits ahead of the last tag the project is.

<H>
 Is the hash of the current version's commit.

Parts between [] are optional.

The translated version is then constructed as:
<X>.<Y>.<Z>[~<pre_release>][+<C>][+dirty.<date>][~<H>][~<metadata>]-<dist>

Where:

<date> is the current date in YYYYmmddHHMMSS format.
<dist> is the current Debian version codename (as returned by lsb_release).

Fixes #117.