Add support for specifying a custom package iteration
Package versioning includes at least two components:

  * the upstream package version (e.g. `1.0.3`)

  * the package iteration (for deb packages, the `debian_revision`)

Makd has until now hardcoded the assumption that the package iteration
should be the same as the distro codename (`trusty`, `xenial`, etc.),
but this is not typical practice in Debian packaging, where it is used
instead to indicate detailed information on the version of the package
(rather than the project being packaged); see:
https://www.debian.org/doc/debian-policy/#version
https://askubuntu.com/questions/620533/how-does-ubuntu-name-packages
for some examples of how the `debian_revision` is set for Debian and
Ubuntu packages respectively.

This patch adds a new, customizable `PKGITERATION` variable which can be
set in build scripts or via the command line.  It defaults to the distro
codename (meaning nothing will change if it is not explicitly set), but
allows the user to override this default with whatever iteration choice
is most appropriate to the package.

The call to `mkpkg` issued from the `$O/pkg-%.stamp` target has been
tweaked to pass in the `PKGITERATION` variable as `iteration`.