Save important generated test files
In case of further debugging is needed, or a re-run of the tests for
other reasons, is good to make sure these generated binaries are kept.
Normally they are built to fulfill the bin.stamp target that runs the
binary, so the binary itself is considered an intermediate files.

For more information, see
https://www.gnu.org/software/make/manual/html_node/Chained-Rules.html#Chained-Rules

Fixes #124.